requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
clearbit==0.1.7
Django==3.1.7
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
idna==2.10
Pillow==8.1.0
psycopg2==2.8.6
pyhunter==1.7
PyJWT==1.7.1
pytz==2021.1
requests==2.25.1
sqlparse==0.4.1
urllib3==1.26.3
